refactor: router component
