+++ src/App.tsx
@@ -1,18 +1,7 @@
 import React from 'react';
-import { createBrowserRouter, Navigate, RouterProvider } from 'react-router-dom';
 import { ThemeProvider, createTheme } from '@mui/material/styles';
 import useMediaQuery from '@mui/material/useMediaQuery';
-import Board from 'routes/board';
-import EditProfile from 'routes/editProfile';
-import ErrorPage from 'routes/errorPage';
-import Main from 'routes/main';
-import Root from 'routes/root';
-import SignIn from 'routes/signIn';
-import Welcome from 'routes/welcome';
-import SignUp from 'routes/signUp';
-import { AppRoutes } from 'types/routes';
-import { useAppSelector } from 'app/hooks';
-import { isUserLoggedIn } from 'components/signForms/authSlice';
+import AppRouter from 'components/router/appRouter';
 
 export const ColorModeContext = React.createContext({ toggleColorMode: () => {} });
 
@@ -27,41 +16,6 @@ export default function App() {
     }),
     []
   );
-  const isLoggedIn = useAppSelector(isUserLoggedIn);
-  const router = createBrowserRouter([
-    {
-      path: AppRoutes.Welcome,
-      element: <Root />,
-      errorElement: <ErrorPage />,
-      children: [
-        {
-          index: true,
-          element: <Welcome />,
-        },
-        {
-          path: AppRoutes.SignIn,
-          element: isLoggedIn ? <Navigate to={AppRoutes.Main} replace /> : <SignIn />,
-        },
-        {
-          path: AppRoutes.SignUp,
-          element: isLoggedIn ? <Navigate to={AppRoutes.Main} replace /> : <SignUp />,
-        },
-        {
-          path: AppRoutes.EditProfile,
-          element: isLoggedIn ? <EditProfile /> : <Navigate to={AppRoutes.Welcome} replace />,
-        },
-        {
-          path: AppRoutes.Main,
-          element: isLoggedIn ? <Main /> : <Navigate to={AppRoutes.Welcome} replace />,
-        },
-        {
-          path: AppRoutes.Board,
-          element: isLoggedIn ? <Board /> : <Navigate to={AppRoutes.Welcome} replace />,
-        },
-      ],
-    },
-  ]);
-
   const theme = React.useMemo(
     () =>
       createTheme({
@@ -75,7 +29,7 @@ export default function App() {
   return (
     <ColorModeContext.Provider value={colorMode}>
       <ThemeProvider theme={theme}>
-        <RouterProvider router={router} />
+        <AppRouter />
       </ThemeProvider>
     </ColorModeContext.Provider>
   );
